Don't burn active probing-resistant PTs.
The GFW used to block bridges by IP address:port but a while ago, it
started to block bridges by IP address instead.  This means that if a
bridge runs obfs4 (which is active probing-resistant) and obfs3 (which
is not active probing-resistant), and BridgeDB happens to hand out the
bridge's obfs3 line, the GFW would manage to probe the bridge, and block
the entire IP address, *including* obfs4.

In this patch, we deal with the GFW's update by not handing out a
bridge's probe-able protocols if it also runs an active
probing-resistant protocol such as obfs4 or scramblesuit.

The patch adds a new configuration option, PROBING_RESISTANT_TRANSPORTS,
which must contain a list of active probing-resistant protocols.

This fixes bug 28655: <https://bugs.torproject.org/28655>