Don't interpret quoted text as commands.
BridgeDB gets confused when users reply to a "get help" email.  The
issue is that BridgeDB interprets commands anywhere in the email body,
even if it's in quoted text.

To fix this issue, we are ignoring commands whose email body line starts
with a '>' character, which is typically used for email quotes.

This fixes <https://bugs.torproject.org/17626>.