Add undocumented OS-level dependency.
BridgeDB uses the binary dkimverify in /home/bridgedb/.procmailrc.  The
binary used to be provided by the python-dkim package (which is
installed on polyanthum) but with the upgrade to Debian Buster, it is
now provided by python3-dkim (which is not installed).  This patch makes
this OS-level dependency explicit.

This commit fixes <https://bugs.torproject.org/32105>.