Fix context manager creation
BridgeDB subclassed a GeneratorContextManager class which became private in
python 3.x. Its constructor signature changed, so attempting to simply use
the private version failed with...

  Traceback (most recent call last):
    File "/home/atagar/Desktop/tor/bridgedb/bridgedb/test/test_Storage.py", line 76, in test_insertBridgeAndGetRing_already_seen_bridge
      with Storage.getDB() as db:
    File "/home/atagar/Desktop/tor/bridgedb/bridgedb/Storage.py", line 400, in helper
      return DBGeneratorContextManager(func(*args, **kwds))
  builtins.TypeError: __init__() missing 2 required positional arguments: 'args' and 'kwds'

Storage.py is pretty confusing, but they seem to simply want a context manager
so simply implementing one.

This changes the test results as follows...

  before: FAILED (skips=106, failures=16, errors=272, successes=410)
  after:  FAILED (skips=106, failures=17, errors=266, successes=415)